vendorsProjects entity moved to domain.layer
