+++ Servers/database/db.ts
@@ -16,7 +16,7 @@ import { TopicModel } from "../domain.layer/models/topic/topic.model";
 import { UserModel } from "../domain.layer/models/user/user.model";
 import { VendorModel } from "../domain.layer/models/vendor/vendor.model";
 import { VendorRiskModel } from "../domain.layer/models/vendorRisk/vendorRisk.model";
-import { VendorsProjectsModel } from "../models/vendorsProjects.model";
+import { VendorsProjectsModel } from "../domain.layer/models/vendorsProjects/vendorsProjects.model";
 import dbConfig from "./config/config";
 import { Dialect } from "sequelize";
 import { FrameworkModel } from "../models/frameworks.model";
